variant tests
